Run a pipeline step-by-step in the debugging browser

debug_pipeline creates a function out of the input pipeline with each line running an additional step from the pipeline and calls the debugging browser on the function. Each line run prints to the console plus stores the result to a variable of the form dot[N], where [N] is the index of the element in the pipeline, so results can easily be replicated or adjusted in the debugging console.

debug_pipeline(pipeline, data = c("pipe", "insert"))

Arguments

pipeline
A pipeline to debug step-by-step Can be an unquoted expression or a character vector of code. If missing, uses the text highlighted in RStudio's source editor.

data
Determines whether data is piped in (data %>% function(parameter)) or inserted as the first parameter of each call (function(data, parameter)). Piping is necessary to make dot notation work; inserting makes stepping into calls easier.

Details

debug_pipeline can also be called interactively in RStudio via the "Debug pipeline in browser" add-in after highlighting the pipeline to debug in the source editor.

For a full explanation of the special commands available in the debugging browser, see browser.

See also

Examples

# NOT RUN {
library(magrittr)

debug_pipeline(
    x <- 1:5 %>% rev %>% {. * 2} %>% sample(replace = TRUE)
)

debug_pipeline(
    x <- 1:5 %>% rev() %>% sample(replace = TRUE),
    data = "insert"
)
# }